+++ consensus/src/main/kotlin/maru/consensus/dummy/EngineApiBlockCreator.kt
@@ -13,10 +13,9 @@
    See the License for the specific language governing permissions and
    limitations under the License.
  */
-package maru.consensus
+package maru.consensus.dummy
 
 import java.util.Optional
-import maru.consensus.dummy.DummyConsensusState
 import maru.core.ExecutionPayload
 import maru.executionlayer.manager.ExecutionLayerManager
 import org.apache.logging.log4j.LogManager
