+++ consensus/src/test/kotlin/maru/consensus/dummy/EngineApiBlockCreatorTest.kt
@@ -13,27 +13,19 @@
    See the License for the specific language governing permissions and
    limitations under the License.
  */
-package maru.consensus
+package maru.consensus.dummy
 
 import java.time.Clock
 import kotlin.random.Random
-import maru.consensus.dummy.DummyConsensusState
-import maru.consensus.dummy.FinalizationState
 import maru.core.ExecutionPayload
-import maru.core.ext.DataGenerators
-import maru.executionlayer.manager.DataGenerators.randomValidForkChoiceUpdatedResult
+import maru.executionlayer.manager.DataGenerators
 import maru.executionlayer.manager.ExecutionLayerManager
 import maru.executionlayer.manager.ForkChoiceUpdatedResult
 import org.hyperledger.besu.ethereum.mainnet.MainnetBlockHeaderFunctions
 import org.junit.jupiter.api.AfterEach
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
-import org.mockito.Mockito.atLeastOnce
-import org.mockito.Mockito.inOrder
-import org.mockito.Mockito.mock
-import org.mockito.Mockito.reset
-import org.mockito.Mockito.verify
-import org.mockito.Mockito.verifyNoMoreInteractions
+import org.mockito.Mockito
 import org.mockito.kotlin.any
 import org.mockito.kotlin.eq
 import org.mockito.kotlin.whenever
@@ -44,19 +36,19 @@ class EngineApiBlockCreatorTest {
 
   @BeforeEach
   fun setUp() {
-    executionLayerManager = mock()
+    executionLayerManager = Mockito.mock()
   }
 
   @AfterEach
   fun tearDown() {
-    reset(executionLayerManager)
+    Mockito.reset(executionLayerManager)
   }
 
   private fun createDummyConsensusState(finalizationState: FinalizationState): DummyConsensusState =
     DummyConsensusState(
       Clock.systemUTC(),
       finalizationState,
-      Random.nextBytes(32),
+      Random.Default.nextBytes(32),
     )
 
   private fun mockFinishBlockBuilding(result: ExecutionPayload) {
@@ -73,9 +65,9 @@ class EngineApiBlockCreatorTest {
 
   @Test
   fun `initialization triggers setHeadAndStartBlockBuilding with latest known state`() {
-    val finalizationState = FinalizationState(Random.nextBytes(32), Random.nextBytes(32))
+    val finalizationState = FinalizationState(Random.Default.nextBytes(32), Random.Default.nextBytes(32))
     val dummyConsensusState = createDummyConsensusState(finalizationState)
-    mockSetHeadAndStartBlockBuilding(randomValidForkChoiceUpdatedResult())
+    mockSetHeadAndStartBlockBuilding(DataGenerators.randomValidForkChoiceUpdatedResult())
     val nextTimestamp = dummyConsensusState.clock.millis()
     EngineApiBlockCreator(
       manager = executionLayerManager,
@@ -84,7 +76,7 @@ class EngineApiBlockCreatorTest {
       initialBlockTimestamp = nextTimestamp,
     )
 
-    verify(executionLayerManager, atLeastOnce()).setHeadAndStartBlockBuilding(
+    Mockito.verify(executionLayerManager, Mockito.atLeastOnce()).setHeadAndStartBlockBuilding(
       eq(dummyConsensusState.latestBlockHash),
       eq(finalizationState.safeBlockHash),
       eq(finalizationState.finalizedBlockHash),
@@ -94,11 +86,13 @@ class EngineApiBlockCreatorTest {
 
   @Test
   fun `block creator doesn't call setHeadAndStartBlockBuilding or finishBlockBuilding twice in a row`() {
-    val finalizationState = FinalizationState(Random.nextBytes(32), Random.nextBytes(32))
+    val finalizationState = FinalizationState(Random.Default.nextBytes(32), Random.Default.nextBytes(32))
     val dummyConsensusState = createDummyConsensusState(finalizationState)
-    val expectedBlockBuildingResult = DataGenerators.randomExecutionPayload()
+    val expectedBlockBuildingResult =
+      maru.core.ext.DataGenerators
+        .randomExecutionPayload()
     mockFinishBlockBuilding(expectedBlockBuildingResult)
-    val forkChoiceUpdatedResult = randomValidForkChoiceUpdatedResult()
+    val forkChoiceUpdatedResult = DataGenerators.randomValidForkChoiceUpdatedResult()
     mockSetHeadAndStartBlockBuilding(forkChoiceUpdatedResult)
     val nextTimestamp = dummyConsensusState.clock.millis()
 
@@ -113,7 +107,7 @@ class EngineApiBlockCreatorTest {
     blockCreator.createEmptyWithdrawalsBlock(2L, null)
     blockCreator.createEmptyWithdrawalsBlock(3L, null)
 
-    val inOrder = inOrder(executionLayerManager)
+    val inOrder = Mockito.inOrder(executionLayerManager)
     repeat((1..3).count()) {
       inOrder.verify(executionLayerManager).setHeadAndStartBlockBuilding(
         any(),
@@ -129,16 +123,18 @@ class EngineApiBlockCreatorTest {
       any(),
       any(),
     )
-    verifyNoMoreInteractions(executionLayerManager)
+    Mockito.verifyNoMoreInteractions(executionLayerManager)
   }
 
   @Test
   fun `finalization updates are respected`() {
-    val finalizationState = FinalizationState(Random.nextBytes(32), Random.nextBytes(32))
+    val finalizationState = FinalizationState(Random.Default.nextBytes(32), Random.Default.nextBytes(32))
     val dummyConsensusState = createDummyConsensusState(finalizationState)
-    val expectedBlockBuildingResult = DataGenerators.randomExecutionPayload()
+    val expectedBlockBuildingResult =
+      maru.core.ext.DataGenerators
+        .randomExecutionPayload()
     mockFinishBlockBuilding(expectedBlockBuildingResult)
-    val forkChoiceUpdatedResult = randomValidForkChoiceUpdatedResult()
+    val forkChoiceUpdatedResult = DataGenerators.randomValidForkChoiceUpdatedResult()
     mockSetHeadAndStartBlockBuilding(forkChoiceUpdatedResult)
     val initialTimestamp = dummyConsensusState.clock.millis()
 
@@ -151,19 +147,19 @@ class EngineApiBlockCreatorTest {
       )
     val nextTimestamp1 = 123L
     blockCreator.createEmptyWithdrawalsBlock(nextTimestamp1, null)
-    verify(executionLayerManager, atLeastOnce()).setHeadAndStartBlockBuilding(
+    Mockito.verify(executionLayerManager, Mockito.atLeastOnce()).setHeadAndStartBlockBuilding(
       eq(expectedBlockBuildingResult.blockHash),
       eq(finalizationState.safeBlockHash),
       eq(finalizationState.finalizedBlockHash),
       eq(nextTimestamp1),
     )
-    val newFinalizationState = FinalizationState(Random.nextBytes(32), Random.nextBytes(32))
+    val newFinalizationState = FinalizationState(Random.Default.nextBytes(32), Random.Default.nextBytes(32))
     dummyConsensusState.updateFinalizationState(newFinalizationState)
 
     val otherTimestamp2 = 124L
     blockCreator.createEmptyWithdrawalsBlock(otherTimestamp2, null)
 
-    verify(executionLayerManager, atLeastOnce()).setHeadAndStartBlockBuilding(
+    Mockito.verify(executionLayerManager, Mockito.atLeastOnce()).setHeadAndStartBlockBuilding(
       eq(expectedBlockBuildingResult.blockHash),
       eq(newFinalizationState.safeBlockHash),
       eq(newFinalizationState.finalizedBlockHash),
